create Dialog folder, rename files, and move
