+++ packages/front-end/components/Experiment/EditDescriptionModal.tsx
@@ -5,7 +5,7 @@ import { ExperimentType } from "shared/validators";
 import { useAuth } from "@/services/auth";
 import Link from "@/ui/Link";
 import MarkdownInput from "@/components/Markdown/MarkdownInput";
-import FormDialog from "@/components/Dialog/FormDialog";
+import DialogLayout from "@/ui/Dialog/Patterns/DialogLayout";
 
 interface Props {
   source: string;
@@ -50,7 +50,7 @@ export default function EditDescriptionModal({
   });
 
   return (
-    <FormDialog
+    <DialogLayout
       trackingEventModalSource={source}
       trackingEventModalType="edit-experiment-description-modal"
       header="Edit Description"
@@ -90,6 +90,6 @@ export default function EditDescriptionModal({
         setValue={(value) => form.setValue("description", value)}
         placeholder={getExperimentDescriptionPlaceholder(experimentType)}
       />
-    </FormDialog>
+    </DialogLayout>
   );
 }
