+++ packages/front-end/components/Experiment/EditPhasesModal.tsx
@@ -5,7 +5,7 @@ import { Box, Flex } from "@radix-ui/themes";
 import { phaseSummary } from "@/services/utils";
 import { useAuth } from "@/services/auth";
 import Button from "@/ui/Button";
-import FormDialog from "@/components/Dialog/FormDialog";
+import DialogLayout from "@/ui/Dialog/Patterns/DialogLayout";
 import Table, {
   TableBody,
   TableCell,
@@ -84,7 +84,7 @@ export default function EditPhasesModal({
     );
   }
   return (
-    <FormDialog
+    <DialogLayout
       trackingEventModalType="edit-phases-modal"
       trackingEventModalSource={source}
       open={true}
@@ -184,6 +184,6 @@ export default function EditPhasesModal({
           New Phase
         </Button>
       )}
-    </FormDialog>
+    </DialogLayout>
   );
 }
